+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Versions Changes
 
+* [v2.1.1](#v211)
+* [v2.1.0](#v210)
+* [v2.0.3](#v203)
+* [v2.0.2](#v202)
+* [v2.0.1](#v201)
+* [v2.0.0](#v200)
 * [v1.2.1](#v121)
 * [v1.2.0](#v120)
 * [v1.1.7](#v117)
+* [v1.1.6](#v116)
 * [v1.1.6](#v116)
 * [v1.1.5](#v115)
 * [v1.1.4](#v114)
@@ -27,16 +34,39 @@
 * [v0.1.0](#v010)
 * [v0.0.5](#v005)
 
+## v2.1.1
+
+The version `v2.1.1` fixes the issue [`#86`](https://github.com/vitalii-andriiovskyi/ngx-owl-carousel-o/issues/86)
+
+## v2.1.0
+
+The version `v2.1.0` changes typings `Event` into `any` in order to avoid the error `ReferenceError: Event is not defined`.
+Also it adds new option `skip_validateItems`.
+
+## v2.0.3
+
+The version `v2.0.3` fixes the bug with `ChangeDetectionStrategy.OnPush`.
+
+## v2.0.2
+
+The version `v2.0.2` adds the option `slideTransition`.
+
+## v2.0.1
+
+The version `v2.0.1` solves the issue [`#38`](https://github.com/vitalii-andriiovskyi/ngx-owl-carousel-o/issues/38).
+
+## v2.0.0
+
+The version `v2.0.0` is generated for Angular 8. It's the same as `v1.1.6` but it's compiled by Angular 8.
+
 ## v1.2.1
 
-The version `v1.2.1` fixes the issue [`#86`](https://github.com/vitalii-andriiovskyi/ngx-owl-carousel-o/issues/86)
+The version `v2.1.1` fixes the issue [`#86`](https://github.com/vitalii-andriiovskyi/ngx-owl-carousel-o/issues/86)
 
 ## v1.2.0
 
 The version `v1.2.0` changes typings `Event` into `any` in order to avoid the error `ReferenceError: Event is not defined`.
-
 Also it adds new option `skip_validateItems`.
-
 
 ## v1.1.7